requirements.txt
azure-identity 
azure-keyvault 
pyodbc 
pandas
requests==2.31.0
fastapi
uvicorn[standard]
pydantic